refactor: modularize project
