+++ controllers/users.js
@@ -1,0 +1,4 @@
+// Get logged-in user's info
+exports.getUserInfo = async (req, res) => {
+    res.json({ status: 'success', user: req.user });
+}
