+++ middleware/logger.js
@@ -1,6 +1,4 @@
-const logger = (req, res, next) => {
+exports.logger = (req, res, next) => {
     console.log(`${req.method} ${req.url}`);
     next();
 };
-
-module.exports = logger;
