+++ middleware/multer.js
@@ -1,0 +1,6 @@
+require('dotenv').config();
+const multer = require('multer');
+const storage = multer.memoryStorage();
+
+// export multer.upload()
+exports.upload = multer({ storage: storage });
