+++ models/bid.js
@@ -1,7 +1,7 @@
 const mongoose = require('mongoose');
 
 const bidSchema = new mongoose.Schema({
-    auction_id: mongoose.Schema.Types.ObjectId,
+    auction_id: mongoose.Schema.Types.ObjectId, // add ref
     user_id: { type: mongoose.Schema.Types.ObjectId, ref: 'User' },
     bid_amount: Number,
     bid_time: Date
