+++ public/bid.html
@@ -100,7 +100,7 @@
         <ul id="bids-list-2"></ul>
     </div>
 
-    <p>copyrigt keter titus 2024</p>
+    <p>copyright keter titus 2024</p>
     <script src="https://cdn.socket.io/4.0.0/socket.io.min.js"></script>
     <script src="js/script.js"></script>
