+++ routes/auth.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+const { register, login } = require('../controllers/auth');
+
+// POST: routes
+router.post('/register', register);
+router.post('/login', login);
+
+module.exports = router;
